change primary key to int, autoeincr.

+++ ums_backend/src/main/java/repository/InternRepository.java
@@ -3,7 +3,7 @@ package repository;
 import models.interns.Intern;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InternRepository extends JpaRepository<Intern, Long>
+public interface InternRepository extends JpaRepository<Intern, Integer>
 {
 
 }
